fix query profile issue of SUBMIT TASK
