+++ fe/fe-core/src/test/java/com/starrocks/qe/ConnectProcessorTest.java
@@ -585,7 +585,7 @@ public class ConnectProcessorTest extends DDLTestBase {
                 ");";
         StatementBase statementBase = UtFrameUtils.parseStmtWithNewParser(sql, ctx);
 
-        processor.addRunningQueryDetail(statementBase);
+        processor.executor.addRunningQueryDetail(statementBase);
 
         Assert.assertFalse(Strings.isNullOrEmpty(QueryDetailQueue.getQueryDetailsAfterTime(0).get(0).getSql()));
     }
